+++ doc/en/CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 - v1.10.6
-  - Add `DeepSeek` and `Gorq` independent Agent support
+  - Add `DeepSeek` and `Groq` independent Agent support
 
 - v1.10.5
   - Add full support for `Telegram MarkdownV2`